requirements.txt
absl-py==2.1.0
altgraph==0.17.4
astunparse==1.6.3
blinker==1.7.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
distlib==0.3.8
dlib @ file:///B:/coding/school%20projects/RESTART/dlib-19.24.1-cp311-cp311-win_amd64.whl#sha256=6f1a5ee167975d7952b28e0ce4495f1d9a77644761cf5720fb66d7c6188ae496
dlib-bin==19.24.2
dm-tree==0.1.8
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.13.1
firebase-admin==6.5.0
Flask==3.0.2
flatbuffers==24.3.7
gast==0.5.4
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==21.2.0
h5py==3.10.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
keras==3.0.5
libclang==16.0.6
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mega.py==1.0.8
ml-dtypes==0.3.2
msgpack==1.0.8
mtcnn==0.1.1
namex==0.0.7
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pathlib==1.0.1
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
Pygments==2.17.2
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PyJWT==2.8.0
pyparsing==3.1.2
pywin32-ctypes==0.2.2
requests==2.31.0
rich==13.7.1
rsa==4.9
six==1.16.0
tenacity==5.1.5
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.1
Werkzeug==3.0.1
wrapt==1.16.0
